Moved helper code from test_obssurface to helpers
This helper code is needed in test_standardise too
