+++ tests/helpers/__init__.py
@@ -24,3 +24,8 @@ from .meta import (
     metadata_checker_obssurface,
     parsed_surface_metachecker,
 )
+from .standardise_results import (
+    filt,
+    select,
+    make_keys,
+)
