+++ tests/store/test_obssurface.py
@@ -10,6 +10,9 @@ from helpers import (
     clear_test_stores,
     get_surface_datapath,
     metadata_checker_obssurface,
+    filt,
+    select,
+    make_keys,
 )
 from openghg.objectstore import (
     exists,
@@ -26,34 +29,6 @@ from openghg.store import ObsSurface
 from openghg.store.base import Datasource
 from openghg.util import create_daterange_str, clean_string
 from pandas import Timestamp
-
-
-def filt(results: list[dict], **kwargs) -> list[dict]:
-    """Filter list of dicts by kwargs"""
-    return [res for res in results if kwargs.items() <= res.items()]
-
-
-def select(results: list[dict], *args) -> list[dict]:
-    """Select key-value pairs from result dicts by keys passed as args"""
-    selected_results = []
-
-    for res in results:
-        selected_results.append({arg: res.get(arg) for arg in args})
-
-    return selected_results
-
-
-def make_keys(results, *args) -> list[str]:
-    """Make strings combining values of metadata."""
-    if not args:
-        args = ("species", "inlet")
-
-    keys = []
-    for res in select(results, *args):
-        values_strings = list(map(str, res.values()))
-        keys.append("_".join(values_strings))
-
-    return keys
 
 
 @pytest.fixture
